requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
rouge-score>=0.1.2
numpy<2.0.0
langchain>=0.1.0
transformers>=4.30.0
torch>=2.0.0
nltk>=3.8.1
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0 